examples/video_shot_classifiers/requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
filelock==3.16.1
fsspec==2024.9.0
huggingface-hub==0.25.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
MarkupSafe==2.1.5
moviepy==1.0.3
mpmath==1.3.0
networkx==3.3
numpy==2.1.1
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
proglog==0.1.10
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scenedetect==0.6.4
sympy==1.13.3
timm==1.0.9
tokenizers==0.19.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
urllib3==2.2.3
